add create routes

+++ database/models/route.py
@@ -2,15 +2,15 @@ from sqlalchemy.orm import relationship
 
 from .base import Base
 
-from sqlalchemy import Column, Integer, String, Enum, Time, Date, ForeignKey, BigInteger
+from sqlalchemy import Column, Integer, String, Enum, Time, Date, ForeignKey, BigInteger, Float
 from sqlalchemy.dialects.postgresql import ARRAY
 
 class Route(Base):
 	__tablename__ = "routes"
 	id = Column(BigInteger, autoincrement=True, primary_key=True)
 	title = Column(String(200), nullable=False)
-	start = Column(ARRAY(BigInteger), nullable=False)
-	stop = Column(ARRAY(BigInteger), nullable=False)
+	start = Column(ARRAY(Float), nullable=False)
+	stop = Column(ARRAY(Float), nullable=False)
 	time_end = Column(Time, nullable=False)
 	time_other = Column(Time, nullable=False, default=0)
 	type_auto = Column(Enum('auto', 'truck', 'taxi', 'walking', name='type_auto'))
